Split skeleton tests and meta-skeleton tests
